chore: refactor to move blocks of functions into thier own files
